+++ src/rxjsOperators/index.ts
@@ -1,2 +1,3 @@
 export * from './handle400WithValidations.operator';
 export * from './handle404.operator';
+export * from './handle4xx.operator';